split brig API
